merge last pr

+++ QLNet/Math/Optimization/LeastSquareProblem.cs
@@ -138,7 +138,6 @@ namespace QLNet
         private double bestAccuracy_;
         //! maximum and real number of iterations
         private int maxIterations_;
-        //private int nbIterations_;
         //! Optimization method
         private OptimizationMethod om_;
         //constraint
@@ -183,11 +182,8 @@ namespace QLNet
             Problem P = new Problem(lsf, c_, initialValue_);
 
             // minimize
-            EndCriteria ec = new EndCriteria(maxIterations_, Math.Min((int)(maxIterations_ / 2), (int)(100)), eps, eps, eps);
+            EndCriteria ec = new EndCriteria(maxIterations_, Math.Min(maxIterations_ / 2, 100), eps, eps, eps);
             exitFlag_ = (int)om_.minimize(P, ec);
-
-            // summarize results of minimization
-            //        nbIterations_ = om_->iterationNumber();
 
             results_ = P.currentValue();
             resnorm_ = P.functionValue();
@@ -225,10 +221,5 @@ namespace QLNet
             return exitFlag_;
         }
 
-        //! return the performed number of iterations
-        //public int iterationsNumber()
-        //{
-        //    return nbIterations_;
-        //}
     }
 }
